move hashing fn to separate file
